Extract row reader to a separate file
