+++ lib/spandx/core/index_file.rb
@@ -2,20 +2,6 @@
 
 module Spandx
   module Core
-    class RowReader
-      attr_reader :io, :index
-
-      def initialize(io, index)
-        @io = io
-        @index = index
-      end
-
-      def row(number)
-        io.seek(index.position_for(number))
-        io.gets
-      end
-    end
-
     class IndexFile
       attr_reader :data_file, :path
